fix(contracts): dedupe Environment types into a shared contract
