+++ frontend/src/lib/types.ts
@@ -3,6 +3,7 @@
  * frontend-only view models used across multiple features.
  */
 import type { RecipeBase, RecipePayload } from "../../../shared/contracts/recipes";
+import type { Environment, EnvironmentEngineId } from "../../../shared/contracts/environments";
 
 // --- Shared contract re-exports ---
 
@@ -86,18 +87,7 @@ export interface RecipeWithStatus extends RecipeBase {
 
 // --- Docker environments (a recipe pinned to an official engine image) ---
 
-export type EnvironmentEngineId = "vllm" | "sglang" | "llamacpp";
-
-export interface Environment {
-  id: string;
-  name: string;
-  recipeId: string;
-  engineId: EnvironmentEngineId;
-  version: string;
-  variant: string | null;
-  createdAt: string;
-  updatedAt: string;
-}
+export type { Environment, EnvironmentEngineId } from "../../../shared/contracts/environments";
 
 export interface EnvironmentWithStatus extends Environment {
   image: string;
