+++ scripts/validate-shared-contracts.mjs
@@ -16,6 +16,8 @@ const contractNames = [
   "ServiceInfo",
   "SystemConfig",
   "EnvironmentInfo",
+  "Environment",
+  "EnvironmentEngineId",
   "RuntimeBackendInfo",
   "EngineBackend",
   "RuntimeKind",
@@ -59,6 +61,7 @@ const allowedFiles = new Set([
   "shared/contracts/controller-events.ts",
   "shared/contracts/observability.ts",
   "shared/contracts/usage.ts",
+  "shared/contracts/environments.ts",
   "controller/src/modules/shared/recipe-types.ts",
   "controller/src/modules/shared/system-types.ts",
   "frontend/src/lib/types.ts",
